Class renamed and refactored
